web/requirements.txt
asn1crypto==1.3.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.2
defusedxml==0.6.0
Django==2.2.20
django-braces==1.14.0
django-extensions==2.2.6
django-filter==2.1.0
django-rest-swagger==2.2.0
djangorestframework==3.11.2
djangorestframework-csv==2.1.0
djangorestframework-gis==0.15
djangorestframework-xml==1.4.0
drf-amsterdam==0.3.1
drf-extensions==0.5.0
factory-boy==2.12.0
Faker==3.0.1
idna==2.8
itypes==1.1.0
MarkupSafe==1.1.1
openapi-codec==1.3.2
psycopg2==2.8.4
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.1
python-keystoneclient==3.22.0
python-swiftclient==3.8.1
pytz==2019.3
requests==2.22.0
sentry-sdk==0.14.2
simplejson==3.17.0
six==1.14.0
text-unidecode==1.3
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.25.8
